+++ docs/changelog.rst
@@ -2,9 +2,9 @@
 Changelog
 =========
 
-* :feature:`-` Clrfund events will now be properly decoded.
 * :bug:`-` Editing the passphrase for OKX exchange will now work properly.
 * :feature:`8193` Coinbasepro support is removed since the exchanges no longer exist.
+* :feature:`8071` Airdrops can now be filtered by a new status: missed. Meaning airdrops for which the user was eligible, the cutoff time has expired and no claim event is seen.
 * :bug:`-` Deposit events in Aave will now show in the correct order.
 * :feature:`3971` Show the total collateral ratio in the Liquity Trove section.
 * :feature:`2323` Add support for Huobi exchange.